# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RLDB - API UNIVERSAL
  description: |
    API de Controle da Rede Logística Digital Brasileira (RLDB) - Integração Crossborder TR8
    
    Esta API funciona como um orquestrador neutro, permitindo que qualquer participante autorizado
    troque dados operacionais sem expor informações competitivas.
    
    A integração TR8 (Transporte Rodoviário Internacional de Cargas) permite operações crossborder
    entre Brasil e outros países, com conformidade total às regulamentações de comércio exterior.
  version: 1.0.0
  contact:
    name: RLDB API Support
    email: api-support@rldb.gov.br
    url: https://rldb.gov.br/api-support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.rldb.gov.br
    description: Production server
  - url: https://sandbox.rldb.gov.br
    description: Sandbox server for testing
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/crossborder/tr8/quotes:
    post:
      tags:
        - TR8 Crossborder
      summary: Get TR8 shipping quotes
      description: Returns available shipping quotes for TR8 international shipments
      operationId: getTR8Quotes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TR8QuoteRequest'
      responses:
        '200':
          description: List of shipping quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrossborderQuote'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/crossborder/tr8/shipments:
    post:
      tags:
        - TR8 Crossborder
      summary: Create TR8 shipment
      description: Creates a new TR8 international shipment
      operationId: createTR8Shipment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TR8ShipmentRequest'
      responses:
        '201':
          description: Shipment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TR8ShipmentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (duplicate idempotency key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/crossborder/tr8/shipments/{shipment_id}:
    get:
      tags:
        - TR8 Crossborder
      summary: Get TR8 shipment
      description: Retrieves a TR8 shipment by its ID
      operationId: getTR8Shipment
      security:
        - bearerAuth: []
      parameters:
        - name: shipment_id
          in: path
          required: true
          schema:
            type: string
          description: The shipment ID
      responses:
        '200':
          description: Shipment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrossborderShipment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Shipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/crossborder/tr8/tracking/{tracking_code}:
    get:
      tags:
        - TR8 Crossborder
      summary: Get TR8 tracking
      description: Retrieves tracking information for a TR8 shipment
      operationId: getTR8Tracking
      security:
        - bearerAuth: []
      parameters:
        - name: tracking_code
          in: path
          required: true
          schema:
            type: string
          description: The tracking code
      responses:
        '200':
          description: Tracking information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TR8TrackingResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tracking code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - TR8 Crossborder
      summary: Update TR8 tracking
      description: Updates tracking information for a TR8 shipment
      operationId: updateTR8Tracking
      security:
        - bearerAuth: []
      parameters:
        - name: tracking_code
          in: path
          required: true
          schema:
            type: string
          description: The tracking code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: New status
                  example: "in_transit"
                location:
                  type: string
                  description: Current location
                  example: "Sao Paulo Hub"
                country:
                  type: string
                  description: Current country
                  example: "BR"
                description:
                  type: string
                  description: Status description
                  example: "Package arrived at hub"
              required:
                - status
                - location
                - country
      responses:
        '200':
          description: Tracking updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tracking code not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/crossborder/tr8/customs/documents:
    post:
      tags:
        - TR8 Crossborder
      summary: Create customs document
      description: Creates a customs document for a TR8 shipment
      operationId: createCustomsDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipment_id:
                  type: string
                  description: The shipment ID
                  example: "TR8-123456789"
                document_type:
                  type: string
                  description: Type of customs document
                  example: "commercial_invoice"
                document_number:
                  type: string
                  description: Document number
                  example: "CI-2024-00001"
                issuing_country:
                  type: string
                  description: Country issuing the document
                  example: "BR"
              required:
                - shipment_id
                - document_type
                - document_number
                - issuing_country
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrossborderCustomsDocument'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/crossborder/tr8/customs/documents/{document_id}:
    get:
      tags:
        - TR8 Crossborder
      summary: Get customs document
      description: Retrieves a customs document by ID
      operationId: getCustomsDocument
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
          description: The document ID
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrossborderCustomsDocument'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/crossborder/tr8/customs/documents/{document_id}/validate:
    put:
      tags:
        - TR8 Crossborder
      summary: Validate customs document
      description: Validates a customs document for a TR8 shipment
      operationId: validateCustomsDocument
      security:
        - bearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
          description: The document ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                verified:
                  type: boolean
                  description: Whether the document is verified
                  example: true
              required:
                - verified
      responses:
        '200':
          description: Document validation updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/crossborder/tr8/webhooks:
    post:
      tags:
        - TR8 Crossborder
      summary: Create webhook
      description: Creates a webhook configuration for TR8 events
      operationId: createWebhook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrossborderWebhook'
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrossborderWebhook'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - TR8 Crossborder
      summary: Get webhooks
      description: Retrieves webhook configurations for the current client
      operationId: getWebhooks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrossborderWebhook'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/observatory/tr8/compliance:
    get:
      tags:
        - Observatory
      summary: Get TR8 compliance report
      description: Returns compliance report for TR8 operations
      operationId: getTR8ComplianceReport
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [last_24_hours, last_7_days, last_30_days, today]
            default: last_30_days
          description: Time period for compliance report
      responses:
        '200':
          description: TR8 compliance report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TR8ComplianceReport'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request"
...
    TR8ShipmentResponse:
      type: object
      properties:
        tracking_code:
          type: string
          description: Tracking code
          example: "TR812345678901"
        label_url:
          type: string
          description: URL to download the shipping label
          example: "https://rldb.gov.br/labels/abc123.pdf"
        estimated_delivery:
          type: string
          format: date-time
          description: Estimated delivery date
        hub_route:
          type: array
          items:
            type: string
          description: Route through hubs
          example: ["Origin Hub", "Destination Hub"]
        shipment_id:
          type: string
          description: Shipment ID
        operator:
          type: string
          description: Operator name
        service:
          type: string
          description: Service name
        customs_status:
          type: string
          description: Customs clearance status
          example: "pending"
        duties_and_taxes_brl:
          type: number
          format: float
          description: Duties and taxes in BRL
        duties_and_taxes_usd:
          type: number
          format: float
          description: Duties and taxes in USD
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
...
security:
  - bearerAuth: []